followed daniel's project structure
