+++ src/modules/user/interfaces.ts
@@ -1,6 +1,5 @@
 import mongoose from "mongoose";
 
-
 export interface IUser {
   _id: string;
   email: string;
