+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ES6",
     "module": "commonjs",
     "rootDir": "src",
-    "outDir": "dist",
+    "outDir": "./dist",
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true
